Fix: 엔티티 타입 변경 (Long -> Integer)

+++ src/main/java/com/boardgo/domain/meeting/entity/MeetingEntity.java
@@ -81,7 +81,7 @@ public class MeetingEntity extends BaseEntity {
 
     @Column
     @ColumnDefault("0")
-    private Long shareCount;
+    private Integer shareCount;
 
     @Builder
     public MeetingEntity(
@@ -101,7 +101,7 @@ public class MeetingEntity extends BaseEntity {
             LocalDateTime meetingDatetime,
             Long hit,
             MeetingState state,
-            Long shareCount) {
+            Integer shareCount) {
         this.id = id;
         this.userId = userId;
         this.title = title;
